+++ .gitignore
@@ -1,2 +1,5 @@
 datasets
 .idea
+wandb
+images
+checkpoints
